XW-1036 | rename storage -> localStorageAdapter
